backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.0.3  # For BaseSettings
langchain-google-genai==1.0.2
browser-use==0.1.0
gtts==2.5.1
aiofiles==23.2.1
python-dotenv==1.0.1
playwright==1.44.0
sqlalchemy==2.0.23        # For database
passlib[bcrypt]==1.7.4    # For password hashing
pyjwt==2.8.0              # For JWT